requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
psd-tools>=1.9.0
